+++ config/Migrations/20201014000225_MigrationPostalCodes.php
@@ -39,7 +39,7 @@ class MigrationPostalCodes extends AbstractMigration {
 				'limit' => 190,
 				'null' => true,
 			])
-			->addColumn('lat', 'float', [
+			->addColumn('lat', 'decimal', [
 				'default' => null,
 				'null' => true,
 				'precision' => 10,
